refactor: 동아리 좋아요 도메인을 하위 패키지로 분리
