+++ src/main/java/in/koreatech/koin/domain/club/like/controller/ClubLikeApi.java
@@ -1,0 +1,53 @@
+package in.koreatech.koin.domain.club.like.controller;
+
+import in.koreatech.koin.global.auth.Auth;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import static in.koreatech.koin.domain.user.model.UserType.STUDENT;
+import static io.swagger.v3.oas.annotations.enums.ParameterIn.PATH;
+
+@Tag(name = "(Normal) Club Like: 동아리 좋아요", description = "동아리에 좋아요를 누르거나 취소한다")
+@RequestMapping("/clubs/{clubId}/like")
+public interface ClubLikeApi {
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "201"),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "동아리 좋아요를 누른다")
+    @PutMapping
+    ResponseEntity<Void> likeClub(
+        @Auth(permit = {STUDENT}) Integer userId,
+        @Parameter(in = PATH) @PathVariable Integer clubId
+    );
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "201"),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "동아리 좋아요를 취소한다")
+    @DeleteMapping("/cancel")
+    ResponseEntity<Void> likeClubCancel(
+        @Auth(permit = {STUDENT}) Integer userId,
+        @Parameter(in = PATH) @PathVariable Integer clubId
+    );
+}
